backend/pyproject.toml
    [tool.poetry]
name = "maddheshiya-studio"
version = "0.1.0"
description = ""
authors = ["Rishikesh Maddhesiya <rishi89324@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.1"
django-cors-headers = "^4.7.0"
asgiref = "^3.8.1"
django-environ = "^0.12.0"
djangorestframework = "^3.16.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.0"
sqlparse = "^0.5.3"
tzdata = "^2025.2"
psycopg = {extras = ["binary"], version = "^3.2.9"}
pillow = "^11.2.1"
djangorestframework-simplejwt = "^5.5.0"
waitress = "^3.0.2"
dj-database-url = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
